requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
psycopg==3.0.7
psycopg2==2.9.3
python-dateutil==1.5
pytz==2021.3
requests==2.27.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
